Implement custom `turbo-offline` element to configure offline mode
This allows users to configure the service worker directly in the HTML
independently of loading Turbo and its configuration, since we can't
change it easily after we've registered the service worker, unlike it
happens with other Turbo configuration options. It'd work like this:

```
<turbo-offline serviceWorkerUrl="/service-worker.js" />
```

And optionally the following attributes can be provided:

- `scope="/"` -> this defaults to resolving "./" against the service
worker's script URL (so, "/" for "/service-worker.js").
- `type="module"`, which can be "classic" or "module", and defaults to "classic".
- `nativeSupport="true"`, which indicates whether the app has a Hotwire
Native counterpart that's loading the service worker as well. We need to
set a cookie in this case, to override the User Agent with Hotwire
Native's custom User Agent, as otherwise the web view's default User
Agent is sent.